spray rudimentaire (augmente de maniere linéaire)

+++ include/structs.h
@@ -85,7 +85,7 @@ struct Player {
   /*Animation de l'arme*/
   int life;
   int cooldown; //nombre d'unités de temps nécéssaires avant de tirer , 0 indique qu'on peut tirer
-  int spray; //nombre d'unités de temps nécéssaires avant de ne plus avoir de spray, 0 indique que le tir sera parfaitement droit
+  double spray; //nombre d'unités de temps nécéssaires avant de ne plus avoir de spray, 0 indique que le tir sera parfaitement droit
   i16 subsector_id;
 
 };
